[SSCP] Ensure custom kernel uses sycl::buffer device pointers
If a kernel passed to `AdaptiveCpp_enqueue_custom_operation` makes use
of (i.e. captures) a `sycl::buffer`, then the captured buffer's internal
pointer must point to its associated device memory, in order for the
custom kernel to extract the pointer and provide to e.g.
`cudaMemcpyAsync`.

The transformation from host pointer to device pointer is done at kernel
launch time by rewriting the bytes of the kernel's captured variables
using an `initialize_embedded_pointers` function. However, this function
was not being called for custom operations enqueued using
`AdaptiveCpp_enqueue_custom_operation`.

So decorate the custom kernel with a call to
`initialize_embedded_pointers`, before calling the kernel itself. Do
the decoration at enqueue time, where we have access to the lambda
object (and its size), rather than trying to introspect the
`std::function` at launch time, with its complex and opaque allocation
behaviour.

Note that there is a test for this under
`tests/sycl/extensions.cpp`:`custom_enqueue`. However, this test is
only conditionally run when targeting `cuda` (or `hip`), and not
for the `generic` target. Hackily enabling this test for the `generic`
target does successfully reproduce the problem, and prove the fix works.
However, enabling the test for the `generic` target requires careful
thought around symbol leakage and/or CI stress, and so is left for
future work (see PR discussion).